Splitting up components into sublibraries
