+++ src/Application/Game.cpp
@@ -33,10 +33,10 @@
 #include "Engine/Graphics/Sprites.h"
 #include "Engine/Graphics/Viewport.h"
 #include "Engine/Graphics/Vis.h"
-#include "Engine/Components/EngineTracer.h"
-#include "Engine/Components/EngineControlComponent.h"
-#include "Engine/Components/EngineDeterministicComponent.h"
-#include "Engine/Components/EngineTraceComponent.h"
+#include "Engine/Components/Trace/EngineTracer.h"
+#include "Engine/Components/Control/EngineControlComponent.h"
+#include "Engine/Components/Deterministic/EngineDeterministicComponent.h"
+#include "Engine/Components/Trace/EngineTraceComponent.h"
 #include "Engine/Localization.h"
 #include "Engine/LOD.h"
 #include "Engine/Objects/Actor.h"
