+++ src/Engine/Components/Trace/CMakeLists.txt
@@ -1,0 +1,23 @@
+cmake_minimum_required(VERSION 3.20.4 FATAL_ERROR)
+
+set(ENGINE_COMPONENTS_TRACE_SOURCES
+        EngineTraceComponent.cpp
+        EngineTracer.cpp)
+
+set(ENGINE_COMPONENTS_TRACE_HEADERS
+        EngineTraceComponent.h
+        EngineTracePlaybackFlags.h
+        EngineTracer.h)
+
+add_library(engine_components_trace STATIC ${ENGINE_COMPONENTS_TRACE_SOURCES} ${ENGINE_COMPONENTS_TRACE_HEADERS})
+target_check_style(engine_components_trace)
+
+target_link_libraries(engine_components_trace
+        engine
+        engine_components_control
+        engine_components_deterministic
+        library_application
+        library_random
+        library_trace
+        platform
+        utility)
