+++ src/Engine/Components/Trace/EngineTraceComponent.cpp
